tools/monthly-report-generator/requirements-backup.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
google-ads==28.4.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
httplib2==0.31.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
